small fixes, rework graph builder from user data
